% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ethanol}
\alias{ethanol}
\title{Measurements of boiling point of ethanol}
\format{
A vector with 9 elements.
}
\usage{
ethanol
}
\description{
It is an artificial data set to demonstrate error in observations.
}
\keyword{datasets}
