% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{hist.norm}
\alias{hist.norm}
\title{Histogram of samples with estimated density function of the normal distribution}
\usage{
hist.norm(x, breaks="Sturges")
}
\arguments{
\item{x}{A vector of samples.}

\item{breaks}{The parameter passed to 'hist'. The default is "Sturges".}
}
\description{
This function draws the histogram given samples with the normal probability density function on the same screen.
   The PDF is estimated from the samples, i.e., 'mean(x)' and 'sd(x)'.
}
\examples{
   x <- rnorm(1000,mean=50,sd=10)
   hist.norm(x)
}
\seealso{
https://stackoverflow.com/questions/2309123/how-to-plot-histogram-and-pdf-together-in-r
}
