% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{hist.t}
\alias{hist.t}
\title{Histogram of samples with estimated density function of t-distribution}
\usage{
hist.t(x, df, breaks="Sturges")
}
\arguments{
\item{x}{A vector of samples.}

\item{df}{Degree of freedom for 'dt', t-distribution.}

\item{breaks}{The parameter passed to 'hist'. The default is "Sturges".}
}
\description{
This function draws the histogram given samples with the probability density function of the t distribution on the same screen.
The user gives the degree of freedom for t-distribution.
}
\seealso{
https://stackoverflow.com/questions/2309123/how-to-plot-histogram-and-pdf-together-in-r
}
