% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{sd.pooled}
\alias{sd.pooled}
\title{Standard deviation for pooled samples from two or more levels}
\usage{
sd.pooled(...)
}
\arguments{
\item{...}{Vectors of samples.}
}
\description{
This function calculates the "pooled" standard deviation required for the t statistics.
   This function is a simple wrapper of 'var.pooled'.
}
