% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{var.pooled}
\alias{var.pooled}
\title{Variance for pooled samples from two or more levels}
\usage{
var.pooled(...)
}
\arguments{
\item{...}{Vectors of samples.}
}
\description{
This function calculates the "pooled" variance required for the t statistics.
}
